requirements.txt
pytz
python-dateutil
GitPython>=0.3.5
boto

